+++ inst/examples/sidebar_app.R
@@ -6,6 +6,7 @@ library(leaflet.extras2)
 data(breweries91, package = "leaflet")
 
 ui <- fluidPage(
+  tags$head(tags$style(".btn-default {display: block;}")),
   h4("Leaflet Sidebar Plugin"),
   splitLayout(cellWidths = c("20%", "80%"),
               tagList(
@@ -42,7 +43,7 @@ ui <- fluidPage(
                     )
                   ),
                   sidebar_pane(
-                    title = "messages", id = "messages_id", icon = icon("person"),
+                    title = "messages", id = "messages_id", icon = icon("person", verify_fa = FALSE),
                     tagList(
                       checkboxGroupInput("variable", "Variables to show:",
                         c("Cylinders" = "cyl",
@@ -69,7 +70,7 @@ server <- function(input, output, session) {
   observe({
     req(input$obs)
     df <- breweries91[sample.int(nrow(breweries91), input$obs), ]
-    bbox <- st_bbox(df)
+    bbox <- suppressWarnings(st_bbox(df))
     leafletProxy("map", session) %>%
       clearGroup("pts") %>%
       addCircleMarkers(data = df,